refactor: move Failure Definition view to view.py
