--- src/ramstk/views/gtk3/failure_definition/listview.pyi
@@ -1,44 +1,0 @@
-# Standard Library Imports
-from typing import Any, Dict
-
-# Third Party Imports
-import treelib
-
-# RAMSTK Package Imports
-from ramstk.configuration import RAMSTKUserConfiguration as RAMSTKUserConfiguration
-from ramstk.logger import RAMSTKLogManager as RAMSTKLogManager
-from ramstk.views.gtk3 import Gtk as Gtk
-from ramstk.views.gtk3 import _ as _
-from ramstk.views.gtk3.widgets import RAMSTKListView as RAMSTKListView
-from ramstk.views.gtk3.widgets import RAMSTKPanel as RAMSTKPanel
-
-class FailureDefinitionPanel(RAMSTKPanel):
-    _module: str = ...
-    _dic_attribute_keys: Any = ...
-    _dic_attribute_updater: Any = ...
-    _dic_row_loader: Any = ...
-    _title: Any = ...
-    def __init__(self) -> None: ...
-    def do_set_callbacks(self) -> None: ...
-    def do_set_properties(self, **kwargs: Dict[str, Any]) -> None: ...
-    def _on_module_switch(self, module: str = ...) -> None: ...
-    _record_id: Any = ...
-    def _on_row_change(self, selection: Gtk.TreeSelection) -> None: ...
-    def __do_load_failure_definition(
-        self, node: treelib.Node, row: Gtk.TreeIter
-    ) -> Gtk.TreeIter: ...
-
-class FailureDefinition(RAMSTKListView):
-    _module: str = ...
-    _tablabel: Any = ...
-    _tabtooltip: Any = ...
-    _lst_mnu_labels: Any = ...
-    _lst_tooltips: Any = ...
-    _pnlPanel: Any = ...
-    def __init__(
-        self, configuration: RAMSTKUserConfiguration, logger: RAMSTKLogManager
-    ) -> None: ...
-    def _do_request_delete(self, __button: Gtk.ToolButton) -> None: ...
-    _record_id: Any = ...
-    def _do_set_record_id(self, attributes: Dict[str, Any]) -> None: ...
-    def __make_ui(self) -> None: ...
